+++ src/middlewared/middlewared/plugins/crypto_/utils.py
@@ -3,6 +3,15 @@ import re
 from cryptography import x509
 
 
+CERT_BACKEND_MAPPINGS = {
+    'common_name': 'common',
+    'country_name': 'country',
+    'state_or_province_name': 'state',
+    'locality_name': 'city',
+    'organization_name': 'organization',
+    'organizational_unit_name': 'organizational_unit',
+    'email_address': 'email'
+}
 # This constant defines the default lifetime of certificate ( https://support.apple.com/en-us/HT211025 )
 DEFAULT_LIFETIME_DAYS = 397
 EC_CURVES = [
